refactor: consistency fixes and boy-scouting
