+++ packages/aio-commerce-lib-app/source/management/association/index.ts
@@ -15,6 +15,6 @@ export {
   clearAssociationData,
   getAssociationData,
   setAssociationData,
-} from "./association-repository";
+} from "./repository";
 
 export type { AssociatedCommerceData, AssociationData } from "./types";
