+++ packages/aio-commerce-lib-app/test/unit/lib/commerce.test.ts
@@ -19,7 +19,7 @@ const { mockGetAssociationData, MockAdobeCommerceHttpClient } = vi.hoisted(
   }),
 );
 
-vi.mock("#management/association/association-repository", () => ({
+vi.mock("#management/association/repository", () => ({
   getAssociationData: mockGetAssociationData,
 }));
 
@@ -27,8 +27,8 @@ vi.mock("@adobe/aio-commerce-lib-api", () => ({
   AdobeCommerceHttpClient: MockAdobeCommerceHttpClient,
 }));
 
-import { getCommerceClient, getCommerceInstance } from "#access/commerce";
-import { AssociationRecordNotFoundError } from "#errors/association-record-not-found-error";
+import { getCommerceClient, getCommerceInstance } from "#lib/commerce";
+import { AssociationRecordNotFoundError } from "#lib/errors";
 
 const auth = { strategy: "ims" } as never;
